- fix for name conflict in proto helpers; - switch to constants defined in m_nudge.h
git-svn-id: http://svn.miranda-ng.org/main/trunk@5504 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SecureIM/src/svcs_proto.cpp
@@ -802,7 +802,7 @@ INT_PTR __cdecl onSendFile(WPARAM wParam, LPARAM lParam)
 			if (!name ) name = file[i];
 			else name++;
 
-			int size = TEMP_SIZE + strlen(name) + 20;
+			int size = TEMP_SIZE + (int)strlen(name) + 20;
 			char *file_out = (char *)mir_alloc(size);
 			mir_snprintf(file_out, size, "%s\\%s.AESHELL(%d)", TEMP, name, file_idx++);
